backend-model-edit/requirements.txt
accelerate==1.8.1
aiohttp>=3.8.5
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4>=4.12.3
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
diffusers==0.34.0
distro==1.9.0
dotenv==0.9.9
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
googlesearch-python>=1.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
huggingface_hub>=0.20.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib-metadata==8.7.0
ip-adapter @ git+https://github.com/tencent-ailab/IP-Adapter.git@62e4af9d0c1ac7d5f8dd386a0ccf2211346af1a2
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
jsonschema-specifications==2025.4.1
jsonschema==4.24.0
jupyter-client==8.6.3
jupyter-core==5.8.1
lazy-loader==0.4
llvmlite==0.44.0
markupsafe==3.0.2
matplotlib
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja
numba==0.61.2
numpy==2.2.6
omegaconf
onnxruntime==1.22.0
openai>=1.93.0
opencv-python-headless==4.11.0.86
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
prompt-toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure-eval==0.2.3
pydantic-core==2.33.2
pydantic==2.11.7
pygments==2.19.2
pymatting==1.1.14
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv
pyyaml==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
rembg==2.0.66
requests>=2.32.4
rpds-py==0.26.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tifffile==2025.5.10
timm
tokenizers==0.21.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
wandb
wcwidth==0.2.13
wheel
xformers==0.0.31
zipp==3.23.0
